Basic Lookback working, multiple tabs, timeline showing up, added TabComponent, yeee bb
